backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
sendgrid==6.11.0
pyairtable==2.2.0
slack-sdk==3.26.1